Try to prevent bridges from arriving after disco is shut down.
Bridges will not get added to the global bridges list after
stop_discovery is called, but if a Bridge#update call is started before
the shutdown but finishes afterward, any bridge callbacks will be
notified that the Bridge has become available, then immediately
unavailable.

Added/updated tests to assist in developing and debugging this change.